requirements.txt
aiogram>=3.4,<4
aiosqlite>=0.19
python-dotenv>=1.0
aiohttp-jinja2>=1.6
jinja2>=3.1
